+++ src/main/java/org/aksw/iguana/cc/tasks/stresstest/metrics/WorkerMetric.java
@@ -1,7 +1,7 @@
 package org.aksw.iguana.cc.tasks.stresstest.metrics;
 
 import org.aksw.iguana.cc.model.QueryExecutionStats;
-import org.aksw.iguana.cc.model.WorkerMetadata;
+import org.aksw.iguana.cc.worker.WorkerMetadata;
 
 import java.util.List;
